requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
async-lru==1.0.2
pyaes==1.6.1
pyrogram[fast]==1.0.7
pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.15.0
tgcrypto==1.2.2
tinydb==4.2.0
